tick systems

+++ Packages/com.fullmetalbagel.overlook-system/ISystemFactory.cs
@@ -7,7 +7,7 @@ public interface ISystemFactory
 {
     string SystemName { get; }
     bool Enable { get; }
-    int TickStage { get; }
+    byte TickStage { get; }
     int TickTimes { get; }
     // TODO: Consider adding ResolveAsync method in the future to support async asset loading
     ISystem Resolve(Container container, int systemIndex);
